Let C backend argument list make a copy of its data
This better matches the semantics of GPU kernels, where arguments are copied to
the GPU. As a side effect, this fixes output verification of kernels with
arguments that are used both for input and output.

The new approach stores the argument data both as a numpy object and a
ctypes wrapper around the numpy object in the repurposed nameddtuple Argument.
This removes the need for the arg_mapping dict.