Turn device interfaces and runner into Python Context Managers
This makes the closing of the device context independent of the
behaviour of the garbage collector. Specifically for a CUDA device, the
garbage collector could delay closing the context until a new context
had already been created, causing unpredictable crashes.

NB: The DeviceInterface, CudaFunctions and SequentialRunner classes no
longer implicitly close their contexts in their destructor, but only in the
__exit__ method now.