Simplify NVML code of controlling the clocks and power limit
The NVML class can operate in two clock-control modes: using application
clocks (default) and using locked clocks. The latter needs to be
explicitly enabled. When one of the two modes is not supported, the
other is automatically used as a fallback. The actual power limit
is read upon initialization of NVML, rather than reading the default
power limit. This value is also restored at the end of the execution.