Remove context manager behaviour of device classes
Instead of explicitly managing the context lifetimes, we now use the device
primary context on all cuda backends. In cupy this is (and was) the default.
In pycuda this is now enabled by the new pycuda function
retain_primary_context.

In the non-cuda backends, the context manager behaviour was not used, and
this therefore does not affect those.

Since pycuda does not wrap cuCtxSetCurrent, we do still have to use the
context stack. Mirroring pycuda.autoinit, we use atexit to pop the last
selected context to prevent pycuda from aborting with an error message
on exit.

Rationale:
This fixes the issue that KernelBuilder objects (in their current form) do
not have an explicitly defined lifetime, and therefore it is not possible to
use a context manager in them. This would lead to unpredictable context
issues if a KernelBuilder destructor would get called at the wrong time.

The new approach always uses the primary context, and therefore there is
no confusion about which context is active when.

A potential downside of dropping the context manager behaviour is that it
makes it harder to add multiple device support for KernelBuilder objects.
However, this was not supported yet anyway. Another potential downside is
reduced robustness when users manage their own cuda contexts in-between
kernel_tuner calls. This should only rarely lead to problems.