Fix evaluation count in PyATF search strategies
The search strategies from PyATF are not guaranteed to always give a unique point. It is possible that
`get_next_coordinates` actually returns the same configurations multiple times. However, these should
not count towards the evaluation limit (`max_feval`).

This commit makes it so that only unique configuration evaluations are counted towards the limit.